feat: React homework (7)
