+++ src/App.tsx
@@ -2,8 +2,10 @@
 import { Suspense } from "react";
 import Spinner from "./components/Spinner";
 import { NavLink, Outlet } from "react-router-dom";
-import styles from "./styles/App.module.css";
+import styles from "./assets/styles/App.module.css";
 import { AuthProvider, useAuth } from "./context/AuthContext";
+import { MantineProvider } from "@mantine/core";
+import { Button, Title } from "@mantine/core";
 
 function Shell() {
   const { isAuthenticated, logout } = useAuth();
@@ -13,14 +15,14 @@ function Shell() {
       <div className={styles.content}>
         {isAuthenticated ? (
           <>
-            <h1>Rick and Morty Wiki</h1>
+            <Title order={1}>Rick and Morty Wiki</Title>
             <nav>
               <ul className={styles.navList}>
                 <li><NavLink to="/" end className={({isActive})=> isActive ? "active" : ""}>Главная</NavLink></li>
                 <li><NavLink to="/characters" className={({isActive})=> isActive ? "active" : ""}>Герои</NavLink></li>
                 <li><NavLink to="/locations"  className={({isActive})=> isActive ? "active" : ""}>Локации</NavLink></li>
                 <li><NavLink to="/episodes"   className={({isActive})=> isActive ? "active" : ""}>Эпизоды</NavLink></li>
-                <li><button onClick={logout} className={styles.logoutBtn}>Выйти</button></li>
+                <li><Button variant="light" color="red" onClick={logout}>Выйти</Button></li>
               </ul>
             </nav>
             <Outlet />
@@ -36,10 +38,12 @@ function Shell() {
 
 export default function App() {
   return (
-    <AuthProvider>
-      <Suspense fallback={<Spinner />}>
-        <Shell />
-      </Suspense>
-    </AuthProvider>
+    <MantineProvider withGlobalStyles withNormalizeCSS> 
+      <AuthProvider>
+        <Suspense fallback={<Spinner />}>
+          <Shell />
+        </Suspense>
+      </AuthProvider>
+    </MantineProvider>
   );
 }
