+++ src/components/Bomb.tsx
