+++ src/types/Character.ts
